+++ x/evm/keeper/state_transition.go
@@ -308,7 +308,10 @@ func (k *Keeper) ApplyMessageWithConfig(
 	// access list preparation is moved from ante handler to here, because it's needed when `ApplyMessage` is called
 	// under contexts where ante handlers are not run, for example `eth_call` and `eth_estimateGas`.
 	if rules := cfg.ChainConfig.Rules(big.NewInt(ctx.BlockHeight()), cfg.ChainConfig.MergeNetsplitBlock != nil); rules.IsBerlin {
-		stateDB.PrepareAccessList(msg.From(), msg.To(), evm.ActivePrecompiles(rules), msg.AccessList())
+		// The access list is prepared without any precompile because it is
+		// filled with only the recipient precompile address in the EVM'hook
+		// call.
+		stateDB.PrepareAccessList(msg.From(), msg.To(), []common.Address{}, msg.AccessList())
 	}
 
 	if contractCreation {